feat: Introduce BigInt support for arbitrary-sized integers
- Updated Value structure to include a new VAL_BIGINT type for handling large integers as decimal strings.
- Modified value creation and arithmetic functions to support BigInt operations, including addition and multiplication.
- Enhanced interpreter evaluation to handle BigInt in expressions and assignments.
- Updated parser to correctly parse integer literals as long long integers.
- Adjusted example files to reflect changes in integer handling.

+++ src/parser/parser.h
@@ -57,7 +57,7 @@ typedef struct ASTNode {
     
     // Değerler (literaller için)
     union {
-        int int_value;
+        long long int_value;
         float float_value;
         char* string_value;
         int bool_value;
